Use hyphens instead of en dashes for unordered lists
Unordered lists in the English version of the book were marked with
en dashes (–), instead of hyphens (-). Those do not produce real
unordered lists, but are treaded in Markdown as a single paragraph.

Conversion was done using the following command:
rg '^–' -g '*.md' --files-with-matches | xargs sed -i 's/^–/-/g'
which uses ripgrep.